Move shortcuts component
